Description:     Create a branch that [initially] contains a copy of the trunk as of r15457. r15457 is the last revision of the trunk to be merged into the metadata_journaling branch, and working with a copy of the trunk that is equivalent to the state in the metadata_journaling branch will make the process of bringing back changes easier.
[This special branch will be deleted when the merge is complete]

+++ test/hyperslab.c
@@ -666,7 +666,7 @@ test_multifill(size_t nx)
 	    sprintf(s, "bad dst[%lu].left", (unsigned long)i);
 	} else if (!DBL_ABS_EQUAL(dst[i].mid, fill.mid)) {
             /* Check if two DOUBLE values are equal.  If their difference
-             * is smaller than the EPSILON value for double, they are
+             * is smaller than the EPSILON value for double, they are 
              * considered equal. See the definition in h5test.h.
              */
 	    sprintf(s, "bad dst[%lu].mid", (unsigned long)i);
@@ -1234,7 +1234,7 @@ main(int argc, char *argv[])
     printf("\n");
 
     /* Set the random # seed */
-    HDsrandom((unsigned)HDtime(NULL));
+    HDsrandom((unsigned long)HDtime(NULL));
 
     /*
      * Open the library explicitly for thread-safe builds, so per-thread
